# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "User - Lendtick",
        "contact": {
            "email": "faujiakbar@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/auth": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authentication",
                "description": "Login Lendtick",
                "operationId": "auth",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "Email for login Lendtick",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password bond to that email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                }
            }
        },
        "/auth/check": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Check Authentication",
                "description": "Check Token Lendtick",
                "operationId": "checkAuth",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Authentication",
                "description": "Refresh Token Lendtick",
                "operationId": "refreshAuth",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/auth/forgot": {
            "put": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Forgot Authentication",
                "description": "Forgot Password",
                "operationId": "forgotAuth",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "nik",
                        "in": "formData",
                        "description": "Using NIK to Forgot Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                }
            }
        },
        "/auth/credentials": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Credential user",
                "description": "Get User Lendtick Credentials",
                "operationId": "credentialsAuth",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID user",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}